+++ src/plugins_types/binary.c
@@ -257,7 +257,7 @@ binary_base64_newlines(char **value, uint32_t *value_size, uint32_t *options, st
 }
 
 static LY_ERR
-lyplg_type_store_binary(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint32_t value_size_bits,
+lyplg_type_store_binary(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint64_t value_size_bits,
         uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data), uint32_t hints,
         const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage, struct lys_glob_unres *UNUSED(unres),
         struct ly_err_item **err)
@@ -409,7 +409,7 @@ lyplg_type_sort_binary(const struct ly_ctx *UNUSED(ctx), const struct lyd_value 
 
 static const void *
 lyplg_type_print_binary(const struct ly_ctx *ctx, const struct lyd_value *value, LY_VALUE_FORMAT format,
-        void *UNUSED(prefix_data), ly_bool *dynamic, uint32_t *value_size_bits)
+        void *UNUSED(prefix_data), ly_bool *dynamic, uint64_t *value_size_bits)
 {
     struct lyd_value_binary *val;
     char *ret;
