+++ src/plugins_types/bits.c
@@ -109,7 +109,7 @@ bits_bit_set(char *bitmap, uint32_t size_bits, uint32_t bit_position)
 }
 
 static void
-lyplg_type_lyb_size_bits(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint32_t *fixed_size_bits)
+lyplg_type_lyb_size_bits(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint64_t *fixed_size_bits)
 {
     const struct lysc_type_bits *type_bits = (struct lysc_type_bits *)type;
 
@@ -280,7 +280,7 @@ bits_items2canon(struct lysc_type_bitenum_item **items, char **canonical)
 }
 
 static LY_ERR
-lyplg_type_store_bits(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint32_t value_size_bits,
+lyplg_type_store_bits(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint64_t value_size_bits,
         uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data), uint32_t hints,
         const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage, struct lys_glob_unres *UNUSED(unres),
         struct ly_err_item **err)
@@ -392,7 +392,7 @@ lyplg_type_sort_bits(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *v
 
 static const void *
 lyplg_type_print_bits(const struct ly_ctx *ctx, const struct lyd_value *value, LY_VALUE_FORMAT format,
-        void *UNUSED(prefix_data), ly_bool *dynamic, uint32_t *value_len_bits)
+        void *UNUSED(prefix_data), ly_bool *dynamic, uint64_t *value_len_bits)
 {
     struct lyd_value_bits *val;
     char *ret;
