+++ src/plugins_types/enumeration.c
@@ -38,7 +38,7 @@
  */
 
 static void
-lyplg_type_lyb_size_enum(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint32_t *fixed_size_bits)
+lyplg_type_lyb_size_enum(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint64_t *fixed_size_bits)
 {
     const struct lysc_type_enum *type_enum = (struct lysc_type_enum *)type;
     LY_ARRAY_COUNT_TYPE u;
@@ -66,7 +66,7 @@ lyplg_type_lyb_size_enum(const struct lysc_type *type, enum lyplg_lyb_size_type 
 }
 
 static LY_ERR
-lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint32_t value_size_bits,
+lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint64_t value_size_bits,
         uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data), uint32_t hints,
         const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage, struct lys_glob_unres *UNUSED(unres),
         struct ly_err_item **err)
@@ -74,7 +74,7 @@ lyplg_type_store_enum(const struct ly_ctx *ctx, const struct lysc_type *type, co
     struct lysc_type_enum *type_enum = (struct lysc_type_enum *)type;
     LY_ERR ret = LY_SUCCESS;
     LY_ARRAY_COUNT_TYPE u;
-    uint32_t fixed_size_bits;
+    uint64_t fixed_size_bits;
     ly_bool found = 0;
     uint32_t value_size;
     int64_t num = 0;
@@ -179,9 +179,9 @@ lyplg_type_sort_enum(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *v
 
 static const void *
 lyplg_type_print_enum(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *value, LY_VALUE_FORMAT format,
-        void *UNUSED(prefix_data), ly_bool *dynamic, uint32_t *value_size_bits)
+        void *UNUSED(prefix_data), ly_bool *dynamic, uint64_t *value_size_bits)
 {
-    uint32_t fixed_size_bits;
+    uint64_t fixed_size_bits;
     int64_t prev_num = 0, num = 0;
     void *buf;
