+++ src/plugins_types/leafref.c
@@ -38,7 +38,7 @@
  */
 
 static void
-lyplg_type_lyb_size_leafref(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint32_t *fixed_size_bits)
+lyplg_type_lyb_size_leafref(const struct lysc_type *type, enum lyplg_lyb_size_type *size_type, uint64_t *fixed_size_bits)
 {
     const struct lysc_type_leafref *type_lr = (struct lysc_type_leafref *)type;
 
@@ -46,7 +46,7 @@ lyplg_type_lyb_size_leafref(const struct lysc_type *type, enum lyplg_lyb_size_ty
 }
 
 static LY_ERR
-lyplg_type_store_leafref(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint32_t value_size_bits,
+lyplg_type_store_leafref(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value, uint64_t value_size_bits,
         uint32_t options, LY_VALUE_FORMAT format, void *prefix_data, uint32_t hints, const struct lysc_node *ctx_node,
         struct lyd_value *storage, struct lys_glob_unres *unres,
         struct ly_err_item **err)
@@ -126,7 +126,7 @@ lyplg_type_sort_leafref(const struct ly_ctx *ctx, const struct lyd_value *val1, 
 
 static const void *
 lyplg_type_print_leafref(const struct ly_ctx *ctx, const struct lyd_value *value, LY_VALUE_FORMAT format,
-        void *prefix_data, ly_bool *dynamic, uint32_t *value_size_bits)
+        void *prefix_data, ly_bool *dynamic, uint64_t *value_size_bits)
 {
     return LYSC_GET_TYPE_PLG(value->realtype->plugin_ref)->print(ctx, value, format, prefix_data, dynamic, value_size_bits);
 }
